pyproject.toml
[project]
name = "sar"
version = "0.1.0"
description = "SAR Airport Dataset utilities and viewer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "albumentations>=2.0.8",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "streamlit>=1.52.1",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
